% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorical.R
\name{cat_var_name}
\alias{cat_var_name}
\title{input is mip_rule, results is character vector with infix names}
\usage{
cat_var_name(x, infix = INFIX_CAT_NAME)
}
\description{
input is mip_rule, results is character vector with infix names
}

