% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocalizer.R
\docType{class}
\name{ErrorLocalizer-class}
\alias{ErrorLocalizer-class}
\title{Base class for class locate errors based on rules and data}
\description{
Base class for class locate errors based on rules and data
}

