% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocate-package.r
\docType{package}
\name{errorlocate-package}
\alias{errorlocate}
\alias{errorlocate-package}
\title{Find errors in data given a set of validation rules.}
\description{
\code{errorlocate} helps identify which cells in a record are likely causing rule
violations.
}
\details{
The package is designed to work with \code{\link[validate:validator]{validate::validator()}} rules. While
\code{validate} can determine whether a record is valid, \code{errorlocate} tries to
identify which fields are most likely erroneous.

This is non-trivial because rules are interdependent: changing one value to
satisfy one rule may violate another rule.

\code{errorlocate} implements record-level error localization based on the
Fellegi-Holt approach. It translates the localization task into a mixed
integer optimization problem and uses a MIP solver to find a minimum-weight
set of fields to change.

Typical workflow:
\itemize{
\item Define rules with \code{\link[validate:validator]{validate::validator()}}.
\item Locate likely errors with \code{\link[=locate_errors]{locate_errors()}}.
\item Replace flagged cells with missing values using \code{\link[=replace_errors]{replace_errors()}}.
}
}
\references{
T. De Waal (2003) Processing of Erroneous and Unsafe Data. PhD thesis, University of Rotterdam.

Van der Loo, M., de Jonge, E, Data Cleaning With Applications in R

E. De Jonge and Van der Loo, M. (2012) Error localization as a mixed-integer program in
editrules.

lp_solve and Kjell Konis. (2011). lpSolveAPI: R Interface for
lp_solve version 5.5.2.0. R package version 5.5.2.0-5.
http://CRAN.R-project.org/package=lpSolveAPI
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-cleaning/errorlocate}
  \item Report bugs at \url{https://github.com/data-cleaning/errorlocate/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/0000-0002-6580-4718}{ORCID})

Authors:
\itemize{
  \item Mark van der Loo \email{mark.vanderloo@gmail.com}
}

}
