% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocation.R
\docType{class}
\name{errorlocation-class}
\alias{errorlocation}
\alias{errorlocation-class}
\title{Error location object}
\description{
Error location contains the result of the error detection
Errors can variable based, and/or record based. A variable based error is a flaw in  uni- or multivariate
distribution. To correct this error multiple observations or the aggregated number would have to be adjusted.
A record based #' errors on the other hand is restricted within one observation.
}
\details{
Current implementation assumes that they are record based: errors are a matrix of rows and columns.
However it is thinkable that errors are purely column based, or dataset based. This
is currently implemented by settings all rows or all values to TRUE.

The information contained in error location object is: \code{values} for each variable is noted if the value is erroneous.
}

