% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocation.R
\docType{class}
\name{errorlocation-class}
\alias{errorlocation-class}
\alias{create_errorlocation}
\title{Error location object}
\description{
\code{errorlocation} contains the result of error localization.
It stores, per cell, whether a value is flagged as erroneous (\code{TRUE}), valid
(\code{FALSE}), or missing (\code{NA}).
\itemize{
\item A record-based error is restricted to one observation.
\code{\link[=errorlocate]{errorlocate()}} using the Fellegi-Holt algorithm assumes errors are
record-based.
\item A variable-based error is a flaw in a uni- or multivariate
distribution. Correcting it typically requires changing multiple observations
or aggregate totals.
}
}
\details{
Current implementation assumes record-based errors.
Retrieve error locations with \code{\link[validate:values]{validate::values()}}, which returns a matrix
with the same dimensions as the checked \code{data.frame}.
For errors that are purely column-based, or dataset-based, \code{errorlocation}
returns a matrix with all
rows or cells set to \code{TRUE}.
\code{\link[validate:values]{validate::values()}} returns \code{NA} for missing values.
}

\section{Fields}{


\itemize{
\item \verb{$errors}: \code{matrix} indicating which values are erroneous (\code{TRUE}),
missing (\code{NA}) or valid (\code{FALSE})
\item \verb{$weight}: The total weight per record. A weight of 0 means no errors were detected.
\item \verb{$status}: The \link[lpSolveAPI:solve.lpExtPtr]{status} of the MIP solver for this record.
\item \verb{$duration}: The number of seconds for processing each record.
}
}

\seealso{
Other error finding: 
\code{\link{errors_removed}()},
\code{\link{expand_weights}()},
\code{\link{locate_errors}()},
\code{\link{replace_errors}()}
}
\concept{error finding}
