% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize-errors.R
\name{errors_as_na}
\alias{errors_as_na}
\title{Set erroneous fields to NA}
\usage{
errors_as_na(data, x, ...)
}
\description{
@param data data to be checked
 @param x \code{\link{validator}} object
}
\details{
Find erronous fields using \code{\link{locate_errors}} and replace these
fields automatically with NA.
}
\seealso{
locate_errors
}

