% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft-rule.R
\name{expect_values}
\alias{expect_values}
\title{expect values}
\usage{
expect_values(values, weights, delta_names = NULL, ...)
}
\arguments{
\item{values}{named list of values.}

\item{weights}{named numeric of equal length as values.}

\item{delta_names}{alternative names for binary variables. (used for log)}

\item{...}{not used}
}
\description{
expect values
}
\keyword{internal}
