% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocalizer.R
\docType{class}
\name{FHLocalizer-class}
\alias{FHLocalizer-class}
\alias{fh_localizer}
\title{Feligi-Holt Errorlocalizer}
\description{
Implementation of the Feligi-Holt algorithm using the \code{ErrorLocalizer} base class.
Given a set of validation rules and a dataset the Feligi-Holt algorithm finds for each record
the smallest (weighted) combination of variables that are erroneous (if any).
}

\note{
Most users do not need this class and can use \code{\link[=locate_errors]{locate_errors()}}.

\code{errorlocalizer} implements feligi holt using a MIP-solver. For problems in which
coefficients of the validation rules or the data are too different, you should consider scaling
the data.
}
