% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mip.R
\name{inspect_mip}
\alias{inspect_mip}
\title{Inspect the MIP problem formulation}
\usage{
inspect_mip(data, x, weight, ...)
}
\arguments{
\item{data}{data to be checked}

\item{x}{validation rules or errorlocalizer object to be used for finding
possible errors.}

\item{weight}{\code{numeric} optional weight specification to be used in the
error localization (see \code{\link[=expand_weights]{expand_weights()}}).}

\item{...}{optional parameters that are passed to \code{\link[lpSolveAPI:lp.control]{lpSolveAPI::lp.control()}} (see details)}
}
\description{
Utility function to inspect the MIP problem for one record.
\code{inspect_mip} can be used as a drop-in replacement for \code{\link[=locate_errors]{locate_errors()}}, but
it only uses the first record when multiple rows are supplied.
}
\details{
This is useful for debugging how one record is translated into a mixed integer
problem, including the generated rules, objective, and LP representation.
See \code{vignette("inspect_mip")} for more details.
}
\examples{
rules <- validator(x > 1)
data <- list(x = 0)
weight <- c(x = 1)

mip <- inspect_mip(data, rules)
print(mip)

# inspect the LP problem (prior to solving it with lpSolveAPI)
lp <- mip$to_lp()
print(lp)

# for large problems write the LP problem to disk for inspection
# lpSolveAPI::write.lp(lp, "my_problem.lp")

# solve the MIP system / find a solution
res <- mip$execute()
names(res)

# lpSolveAPI status of finding a solution
res$s

# LP problem after solving (often simplified version of first LP)
res$lp

# records that are deemed "faulty"
res$errors

# values of variables used in the MIP formulation. Also contains a valid solution
# for "faulty" variables
res$values

# see the derived MIP rules and objective function, used in the construction of
# the LP problem
mip$mip_rules()
mip$objective
}
\seealso{
Other Mixed Integer Problem: 
\code{\link{MipRules-class}}
}
\concept{Mixed Integer Problem}
