% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorical.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if rules are categorical}
\usage{
is_categorical(x, ...)
}
\arguments{
\item{x}{validator object}

\item{...}{not used}
}
\value{
logical indicating which rules are purely categorical/logical
}
\description{
Check if rules are categorical
}

