% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.r
\name{is_linear}
\alias{is_linear}
\title{Check which rules are linear rules}
\usage{
is_linear(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object containing data validation rules}
}
\description{
Check which rules are linear rules
}

