% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MipRules.R
\docType{class}
\name{MipRules-class}
\alias{MipRules-class}
\alias{miprules}
\title{Create a mip object from a validator object}
\description{
Create a \code{MipRules} object from \code{\link[validate:validator]{validate::validator()}} rules.
This utility class translates rules into a mixed integer problem.
}
\details{
Most users should use \code{\link[=locate_errors]{locate_errors()}}, which handles translation and
execution automatically. \code{MipRules} is mainly for advanced users who want to
inspect or customize the optimization setup.
}

\section{Methods}{

The \code{MipRules} class contains the following methods:
\itemize{
\item \verb{$execute()} solves the mixed integer problem.
\item \verb{$to_lp()} transforms the object into an \code{lp_solve} problem object.
\item \verb{$is_infeasible} checks whether the current rule system is infeasible.
\item \verb{$set_values()} sets observed values and weights (objective function).
}
}

\examples{
rules <- validator(x > 1)
mr <- miprules(rules)
mr$to_lp()
mr$set_values(c(x=0), weights=c(x=1))
mr$execute()
}
\seealso{
Other Mixed Integer Problem: 
\code{\link{inspect_mip}()}
}
\concept{Mixed Integer Problem}
