% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize-errors.R
\name{replace_errors}
\alias{replace_errors}
\title{Replace erroneous fields with NA or suggested value}
\usage{
replace_errors(data, x, value = c("NA", "suggestion"), ...)
}
\description{
@param data data to be checked
 @param x \code{\link{validator}} object
 @param value \code{NA}
}
\details{
Find erronous fields using \code{\link{locate_errors}} and replace these
fields automatically with NA.
}
\seealso{
locate_errors
}

