% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_constraint.R
\name{set_bounds_from_constraints}
\alias{set_bounds_from_constraints}
\title{extract boundary conditions from linear constraints}
\usage{
set_bounds_from_constraints(x)
}
\arguments{
\item{x}{\verb{[OP]} with linear constraints}

\item{types}{named \code{character} with types of variables}
}
\description{
extract boundary conditions from linear constraints
}
