% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soft-edits.R
\name{soft_edits}
\alias{soft_edits}
\title{Derive editmatrix with soft constraints based on boundaries of variables. This is a utility function that is used for
constructing a mip/lp problem.}
\usage{
soft_edits(E, prefix = "delta.", ...)
}
\arguments{
\item{E}{normalized \code{editmatrix}}

\item{prefix}{\code{character} used for naming dummy variables in matrix.}
}
\description{
Derive editmatrix with soft constraints based on boundaries of variables. This is a utility function that is used for
constructing a mip/lp problem.
}

