% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{substitute_}
\alias{substitute_}
\title{substitute an existing language object}
\usage{
substitute_(x, values = list())
}
\arguments{
\item{x}{expression or language object!}

\item{values}{list of values}
}
\description{
substitute an existing language object
}
