% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocation.R
\name{summary.errorlocation}
\alias{summary.errorlocation}
\title{Summary of errorlocation object}
\usage{
summary.errorlocation(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary of errorlocation
}

