% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mip-interface.R
\name{translate_lp}
\alias{translate_lp}
\title{translate linear rules into an lp problem}
\usage{
translate_lp(rules, objective = NULL, eps = 1e-07, bin_vars = integer())
}
\description{
translate linear rules into an lp problem
}

