% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip_lpsolve.R
\name{translate_mip_lp}
\alias{translate_mip_lp}
\title{Translate linear rules into an LP problem}
\usage{
translate_mip_lp(rules, objective = NULL, eps = 0.001, ...)
}
\arguments{
\item{rules}{MIP rules}

\item{objective}{function}

\item{eps}{accuracy for equality/inequality}

\item{...}{additional \code{\link[lpSolveAPI:lp.control]{lpSolveAPI::lp.control()}} parameters that are set for the MIP problem}
}
\description{
Translate linear rules into an LP problem
}
