% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_mip_OP.R
\name{translate_mip_OP}
\alias{translate_mip_OP}
\title{Translates the problem into an ROI::OP mip problem}
\usage{
translate_mip_OP(rules, objective = NULL, eps = 0.001, ...)
}
\arguments{
\item{rules}{mip rules}

\item{objective}{function}

\item{eps}{accuracy for equality/inequality}

\item{...additional}{parameters that are set for the mip problem}
}
\description{
Translates the problem into an ROI::OP mip problem
}
