% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip_ROIsolve.R
\name{translate_mip_ROI}
\alias{translate_mip_ROI}
\title{translate linear rules into an lp problem}
\usage{
translate_mip_ROI(rules, objective = NULL, eps = 0.001, ...)
}
\arguments{
\item{rules}{mip rules}

\item{objective}{function}

\item{eps}{accuracy for equality/inequality}

\item{...}{additional \code{\link[=lp.control]{lp.control()}} parameters that are set for the mip problem}
}
\description{
translate linear rules into an lp problem
}
