% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cat_as_mip_rules}
\alias{cat_as_mip_rules}
\title{get categorical rules as mip_rules}
\usage{
cat_as_mip_rules(x, ...)
}
\arguments{
\item{x}{expressionset object}

\item{...}{not used}
}
\value{
list of mip_rule
}
\description{
get categorical rules as mip_rules
}
\keyword{internal}
