% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cat_coefficients}
\alias{cat_coefficients}
\title{Get coefficient matrix from categorical rules}
\usage{
cat_coefficients(x, ...)
}
\arguments{
\item{x}{validator object}

\item{...}{not used}
}
\description{
Get coefficient matrix from categorical edits, similar to
linear_coefficients.
}
\keyword{internal}
