% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocalizer.R
\docType{class}
\name{ErrorLocalizer-class}
\alias{ErrorLocalizer-class}
\title{Base class for class locate errors based on rules and data}
\description{
ErrorLocalizer can be used as a base class to implement a new error localization algorithm.
The derived class must implement two methods: \code{initialize}, which is called
before any error localization is done and \code{locate} which operates upon data. The
extra parameter \code{...} can used to supply algorithmic specific parameters.
}

