% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorlocation.R
\docType{class}
\name{errorlocation-class}
\alias{errorlocation}
\alias{errorlocation-class}
\title{Error location object}
\description{
Error location contains the result of the error detection.
Errors can record based or variable based.
\itemize{
\item A record based error is restricted within one observation.
\code{\link{errorlocate}} using the Felligi Holt algorithm assumes errors are record based.
\item A variable based error is a flaw in  uni- or multivariate
 distribution. To correct this error multiple observations or the aggregated number should be adjusted.
}
}
\details{
Current implementation assumes that errors are record based. The error locations can be retrieved
using the method \code{values} and are a matrix of
rows and columns, with the same dimensions are the \code{data.frame} that was checked.
For errors that are purely column based, or dataset based, errorlocations will return a matrix with all
rows or cells set to \code{TRUE}.
The \code{values} return \code{NA} for values that are

The information contained in error location object is: \code{values} for each variable is noted if the value is erroneous.
}

