% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorlocation.R
\docType{class}
\name{errorlocation-class}
\alias{errorlocation}
\alias{errorlocation-class}
\title{Error location}
\description{
Errors can variable based, and/or record based
Current implementation assumes that it  is record based: errors are a matrix of rows and columns.
However it is thinkable that errors are purely column based, or dataset based. This
is currently implemented by settings all rows or all values to TRUE.
}

