% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localize-errors.R
\name{errors_as_na}
\alias{errors_as_na}
\title{utility function that replaces errors with NA}
\usage{
errors_as_na(data, x, ref = NULL, ...)
}
\description{
utility function that replaces errors with NA
}

