% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_mip.R
\name{inspect_mip}
\alias{inspect_mip}
\title{inspect the mip problem for a record}
\usage{
inspect_mip(data, rules, weight)
}
\arguments{
\item{data}{data to be checked}

\item{weight}{\code{numeric} optional weight vector to be used in the error localization.}
}
\description{
Utility function to inspect the mip problem for a record
}
\examples{
rules <- validator(x > 1)
data <- list(x = 0)

mip <- inspect_mip(data, rules)
mip$rules
mip$mip_rules()
mip$objective
mip$to_lp()

res <- mip$execute()
res
res$lp
}
