% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if rules are categorical}
\usage{
is_categorical(x, ...)
}
\arguments{
\item{x}{validator or expression object}

\item{...}{not used}
}
\value{
logical indicating which rules are purely categorical/logical
}
\description{
Check if rules are categorical
}
\details{
#' @note \code{errorlocate} supports linear,
categorical and conditional rules to be used in finding errors. Other rule types
are ignored during error finding.
}
\examples{
v <- validator( A \%in\% c("a1", "a2")
              , B \%in\% c("b1", "b2")
              , if (A == "a1") B == "b1"
              , y > x
              )

is_categorical(v)
}
\seealso{
Other rule type: 
\code{\link{is_conditional}()},
\code{\link{is_linear}()}
}
\concept{rule type}
