% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditional.R
\name{is_conditional}
\alias{is_conditional}
\title{Check if rules are conditional rules}
\usage{
is_conditional(rules, ...)
}
\arguments{
\item{rules}{validator object}
}
\value{
logical indicating which rules are conditional
}
\description{
Check if rules are conditional rules
}

