% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{is_linear}
\alias{is_linear}
\title{Check which rules are linear rules.}
\usage{
is_linear(x, ...)
}
\arguments{
\item{x}{\code{\link[=validator]{validator()}} object containing data validation rules}

\item{...}{not used}
}
\value{
\code{logical} indicating which rules are (purely) linear.
}
\description{
Check which rules are linear rules.
}
\note{
\code{errorlocate} supports linear,
categorical and conditional rules to be used in finding errors. Other rule types
are ignored during error finding.
}
\seealso{
Other rule type: 
\code{\link{is_categorical}()},
\code{\link{is_conditional}()}
}
\concept{rule type}
