% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize-errors.R
\docType{methods}
\name{locate_errors}
\alias{locate_errors}
\alias{locate_errors,data.frame,ErrorLocalizer-method}
\alias{locate_errors,data.frame,validator-method}
\title{Locate errors in data}
\usage{
locate_errors(data, x, ...)

\S4method{locate_errors}{data.frame,validator}(data, x, weight = NULL,
  ref = NULL, ...)

\S4method{locate_errors}{data.frame,ErrorLocalizer}(data, x, weight = NULL,
  ref = NULL, ...)
}
\arguments{
\item{data}{data to be checked}

\item{x}{validation rules or errorlocalizer object to be used for finding
possible errors.}

\item{...}{optional parameter to be used by a specific method}

\item{weight}{\code{numeric} optional weight vector to be used in the error localization.}

\item{ref}{\code{data.frame} optional reference data to be used in the rules checking}
}
\value{
\code{\link{errorlocation}} object describing the errors found.

Locate erronuous fields in rows of data using validation rules or a specific
errorlocalizer object. This method returns the errors, according to the specified
method \code{x}, that were found. If these errors are to be removed automatically
}
\description{
Locate errors in data
}
\examples{
v <- validator( profit + cost == turnover
              , cost - 0.6*turnover >= 0
              , cost>= 0
              , profit >= 0
)
data <- data.frame(profit=755, cost=125, turnover=200)
locate_errors(data, v)


# try a condinational rule
v <- validator( maried \%in\% c(TRUE, FALSE), if (maried==TRUE) age >= 17 )
data <- data.frame( married = TRUE, age = 16)
locate_errors(data, v)

}
\seealso{
locate_errors
}

