% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip_rule.R
\name{mip_rule}
\alias{mip_rule}
\title{Create a rule used by mip}
\usage{
mip_rule(a, op, b, rule, type, weight = Inf, ...)
}
\arguments{
\item{a}{named vector with coefficients}

\item{op}{operator in ("<=", "==", ">=", ">", "<")}
}
\description{
Create a rule used by mip
}
\keyword{internal}
