% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize-errors.R
\docType{methods}
\name{replace_errors}
\alias{replace_errors}
\alias{replace_errors,data.frame,ErrorLocalizer-method}
\alias{replace_errors,data.frame,errorlocation-method}
\alias{replace_errors,data.frame,validator-method}
\title{Replace erroneous fields with NA or suggested value}
\usage{
replace_errors(data, x, ref = NULL, ..., value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,validator}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,ErrorLocalizer}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,errorlocation}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))
}
\arguments{
\item{data}{data to be checked}

\item{x}{\code{\link{validator}} object}

\item{ref}{optional reference data set}

\item{...}{other datasets to be used...}

\item{value}{\code{NA}}
}
\description{
Find erronous fields using \code{\link{locate_errors}} and replace these
fields automatically with NA.
}
\seealso{
locate_errors
}

