% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-errors.R
\docType{methods}
\name{replace_errors}
\alias{replace_errors}
\alias{replace_errors,data.frame,ErrorLocalizer-method}
\alias{replace_errors,data.frame,errorlocation-method}
\alias{replace_errors,data.frame,validator-method}
\title{Replace erroneous fields with NA or a suggested value}
\usage{
replace_errors(data, x, ref = NULL, ..., value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,validator}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,ErrorLocalizer}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))

\S4method{replace_errors}{data.frame,errorlocation}(data, x, ref = NULL, ...,
  value = c("NA", "suggestion"))
}
\arguments{
\item{data}{data to be checked}

\item{x}{\code{\link{validator}} object}

\item{ref}{optional reference data set}

\item{...}{other datasets to be used...}

\item{value}{\code{NA}}
}
\description{
Find erronous fields using \code{\link{locate_errors}} and replace these
fields automatically with NA. The error location algorithm may provide a suggestion which
validates the record against the rules.
}
\note{
In general it is better to replace the erronuous fields with \code{NA} and apply a proper
imputation methods. Suggested values from the error localization method may introduce an unwanted bias.
}
\examples{
rules <- validator( profit + cost == turnover
              , cost - 0.6*turnover >= 0
              , cost>= 0
              , profit >= 0
)
data <- data.frame(profit=755, cost=125, turnover=200)

data_no_error <-
  data \%>\%
  replace_errors(rules)
}
\seealso{
locate_errors
}

