% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip_lpsolve.R
\name{translate_mip_lp}
\alias{translate_mip_lp}
\title{translate linear rules into an lp problem}
\usage{
translate_mip_lp(rules, objective = NULL, eps = 0.001)
}
\arguments{
\item{rules}{mip rules}

\item{objective}{function}

\item{eps}{accuracy for equality/inequality}
}
\description{
translate linear rules into an lp problem
}
